+++ salon-tablet.code-workspace
@@ -32,7 +32,8 @@
           "msapplication",
           "myproject",
           "noindex",
-          "qrcode"
+          "qrcode",
+          "samplefooter"
         ]
     }
 }
