+++ src/app/View/Footer/Footer/context/useDeleteFooterItem.ts
@@ -1,0 +1,6 @@
+import { useDeleteFooterItem as useDeleteFooterItemAction } from "../../../../ActionCreator/footerItems/useDeleteFooterItem";
+
+export const useDeleteFooterItem = () => {
+  const deleteFooterItem = useDeleteFooterItemAction();
+
+}
